update source code for this module

+++ Controller/Adminhtml/Faq/Delete.php
@@ -5,10 +5,12 @@
  * @Author              Ngo Quang Cuong <bestearnmoney87@gmail.com>
  * @Date                2016-12-18 02:25:15
  * @Last modified by:   nquangcuong
- * @Last Modified time: 2016-12-18 02:27:45
+ * @Last Modified time: 2016-12-20 21:19:56
  */
 
 namespace PHPCuong\Faq\Controller\Adminhtml\Faq;
+
+use PHPCuong\Faq\Model\ResourceModel\Faq;
 
 class Delete extends \Magento\Backend\App\Action
 {
@@ -34,10 +36,28 @@ class Delete extends \Magento\Backend\App\Action
                 // init model and delete
                 $model = $this->_objectManager->create('PHPCuong\Faq\Model\Faq');
                 $model->load($faq_id);
-                $title = $model->getTitle();
-                $model->delete();
-                $this->messageManager->addSuccess(__('The '.$title.' FAQ has been deleted.'));
-                return $resultRedirect->setPath('*/*/');
+                if ($model->getFaqId()) {
+
+                    $title = $model->getTitle();
+
+                    $model->delete();
+
+                    $url_rewrite_model = $this->_objectManager->create('Magento\UrlRewrite\Model\UrlRewrite');
+
+                    $urls_rewrite = $url_rewrite_model->getCollection()
+                    ->addFieldToFilter('entity_type', Faq::FAQ_ENTITY_TYPE)
+                    ->addFieldToFilter('entity_id', $faq_id)
+                    ->load()->getData();
+
+                    foreach ($urls_rewrite as $value) {
+                        $url_rewrite_model = $this->_objectManager->create('Magento\UrlRewrite\Model\UrlRewrite');
+                        $url_rewrite_model->load($value['url_rewrite_id'])->delete();
+                    }
+
+                    $this->messageManager->addSuccess(__('The '.$title.' FAQ has been deleted.'));
+                    return $resultRedirect->setPath('*/*/');
+                }
+
             } catch (\Exception $e) {
                 // display error message
                 $this->messageManager->addError($e->getMessage());
